+++ lib/features/authentication/screens/onboarding/widgets/onboarding_page.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:mystore/utils/constants/sizes.dart';
+import 'package:mystore/utils/helpers/helper_functions.dart';
+
+class OnBoardingPage extends StatelessWidget {
+  const OnBoardingPage({
+    super.key,
+    required this.image,
+    required this.title,
+    required this.subtitle,
+  });
+
+  final String image, title, subtitle;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(MySizes.defaultSpace),
+      child: Column(
+        children: [
+          Image(
+              width: MyHelperFunctions.screenWidth(context) * 0.8,
+              height: MyHelperFunctions.screenHeight(context) * 0.6,
+              image: AssetImage(image)),
+          Text(
+            title,
+            style: Theme.of(context).textTheme.headlineMedium,
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(
+            height: MySizes.spaceBtwItems,
+          ),
+          Text(
+            subtitle,
+            style: Theme.of(context).textTheme.bodyMedium,
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+}
