+++ lib/features/authentication/screens/onboarding/widgets/onboarding_skip.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:mystore/utils/constants/sizes.dart';
+import 'package:mystore/utils/device/device_utility.dart';
+
+class OnBoardingSkip extends StatelessWidget {
+  const OnBoardingSkip({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Positioned(
+      top: MyDeviceUtils.getAppBarHeight(),
+      right: MySizes.defaultSpace,
+      child: TextButton(
+        onPressed: () {},
+        child: const Text('Skip'),
+      ),
+    );
+  }
+}
